requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-multipart==0.0.9
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf==4.2.0
python-dotenv==1.0.0
openai
tiktoken